When getState stops being called(browser throttling)show modal to continue
Browsers will restrict chained calls after 5 minutes of a tab being not
visible. At this point the calls will only be made once per minute. This
would mean that our coach is removed from the backend, only to reappear
every minute and disappear again.

Now, if we detect that a gap of more than 10 seconds has occurred between
calls to getState, we end the chain so that getState is no longer called
at all and we show a modal asking if they want to continue. If they click
on continue we call activate and we begin the chained getState calls again
and hide the modal.