Improve UI responsiveness when hiding matches
Hold a queue of hidden matches.
Keep hidden matches in the queue for 60 seconds.
Use the queue to determine when to fade out and when to hide altogether.
Store date hidden in the queued object to allow fine control.
When processing queue, delay 300ms before emitting main event, allow CSS
transitions to finish.